perf(bundle): lazy-load DOCX/PPTX/XLSX engines and heavy surfaces (P1.3)
Cuts the main startup chunk ~38% raw / ~36% gzip (5,301kB -> 3,293kB raw;
1,609kB -> 1,027kB gzip), measured with vite build before/after.

- DOCX/PPTX/XLSX extraction and export are now dynamically imported on
  first use instead of at startup (ai-file-context.ts, MainPanel.tsx,
  FormattingToolbar.tsx, useDocumentCreation.ts).
- Root-caused two hidden static-import leaks that were forcing
  mammoth/docx-preview/jszip/docx and xlsx back into the main chunk despite
  the above: docx-commands.ts and flushDirtyTabs.ts each statically
  imported one pure-JS helper from the heavy modules. Extracted those
  helpers into new zero-dependency modules (docx-table-utils.ts, and
  dataUrlToArrayBuffer moved into file-utils.ts).
- React.lazy for Email, Activity Log, Privacy Center, and full-page
  Settings (AppSurfaceRouter.tsx); the Account window / connector setup
  panels (AppDialogs.tsx); and the 9 connector citation-viewer panels
  bundled into one lazy chunk (ConnectorSourcePanels.tsx). Client Map,
  Ask, and Documents stay eager (primary demo path, no Suspense flash).
- useOpenFileAIContext.ts extraction now scheduled via
  requestIdleCallback (setTimeout fallback) instead of running inline.

Verified: typecheck clean, full Vitest suite green (one pre-existing
unrelated OCR-fixture failure, unchanged before/after), sharded
Playwright e2e suite (236 tests) green on two separate clean runs.
