+++ CHANGELOG.md
@@ -31,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **Faster first paint: main startup bundle cut ~38% (raw) / ~36% (gzip) by lazy-loading everything not needed on first screen (P1.3).** Measured with `npx vite build` on identical code before/after: main chunk 5,301 kB → 3,293 kB raw (1,609 kB → 1,027 kB gzip). Changes: (1) DOCX/PPTX/XLSX extraction and export are now `await import()`'d on first use instead of imported at startup — `ai-file-context.ts`, `MainPanel.tsx`, `FormattingToolbar.tsx`, `useDocumentCreation.ts`; (2) the pure-JS Markdown-table-detection helpers used by the AI redliner were split out of `docx-io.ts` into a new zero-dependency `docx-table-utils.ts`, and the base64 `dataUrlToArrayBuffer` helper moved from `spreadsheet-io.ts` into the already-zero-dependency `file-utils.ts` — both were previously forcing the full `mammoth`/`docx-preview`/`jszip`/`docx` and `xlsx` libraries into the startup bundle via a single always-on import (`docx-commands.ts`, `flushDirtyTabs.ts`) even though only a tiny pure function was needed; (3) Email, Activity Log, Privacy Center, and the full-page Settings surface are now `React.lazy` in `AppSurfaceRouter.tsx`; the Account window (which pulls in every connector "Connect" setup panel) is lazy in `AppDialogs.tsx`; and the 9 connector citation-viewer panels (Wealthbox, OneDrive, DocuSign, Calendly, Box, ShareFile, Jotform, Zocks, Addepar) are bundled into one lazy chunk (`ConnectorSourcePanels.tsx`) instead of 9 always-on imports in `App.tsx`; (4) open-file AI-context extraction (`useOpenFileAIContext.ts`) is now scheduled via `requestIdleCallback` (falls back to `setTimeout` where unsupported) so it never competes with typing/rendering. Client Map, Ask, and Documents stay eagerly loaded — no Suspense flash on the primary demo path. Verified with a full `npm run typecheck`, the full Vitest suite (5,065 tests, one pre-existing unrelated OCR-fixture failure), and two clean rounds of the sharded Playwright e2e suite (236 tests across 6 shards, both green).
+
 ### Fixed
 - **The AI file-tool workspace-boundary guards are now real cross-platform checks instead of a no-op (F2.8 — closes the path-guard deferral flagged by F2.2).** Every AI file tool in the chat send path (`read_file` / `list_files` / `search_files` / `write_file` / `create_folder` / `move_file` / `delete_file`) built its absolute path with a hand-rolled `` `${rootPath}/${rel}`.replace(/\/+/g,'/') `` template and then gated it with a raw `filePath.startsWith(rootPath)` check. That guard was a **tautology** — the joined path is literally `rootPath + "/" + rel`, so it ALWAYS starts with `rootPath` and never actually rejected anything; the real workspace boundary was only enforced downstream by `PathValidator`. It also could not be turned into a genuine check by normalizing the join alone, because on Windows `rootPath` carries native backslashes (forward-slash-normalizing just the join would make `startsWith` reject **every** legitimate Windows path — fail-closed). The join + guard are now migrated **together** at all 8 sites: joins → `workspacePath(rootPath, rel)` (the shared helper: rejects non-string args, passes an already-absolute `rel` through instead of doubling it, returns a forward-slash-normalized path) and workspace-boundary checks → `sameOrInside(rootPath, filePath)` (segment-boundary, separator-, and volume-root-case-correct). The matter-boundary guards (`assertInActiveMatter` / `assertDirInActiveMatter` / `pathInActiveMatter`, which already use `sameOrInside` via `resolveMatterId`) and `moveToTrash` now receive a cleanly-normalized absolute path on every platform. **Net security effect:** an absolute path pointing OUTSIDE the workspace is now genuinely refused at the tool layer (the old tautology let it reach `PathValidator` before being caught — not an exploitable escape, since `PathValidator` backstopped it, but a dead guard that gave false confidence and, for an in-workspace absolute path, silently produced a doubled/garbage path); a sibling root that merely shares a string prefix is refused (segment boundary); a case-drifted absolute path fails CLOSED on Windows (the intended isolation-safe direction, consistent with `appPath`'s CASE POLICY and `resolveMatterId`'s ambiguity guard); and `..` traversal is still refused by the matter guard in every scope. Behavior for a normal relative path is unchanged (it joins onto the real root, preserving its exact case). Files: `src/features/ask/hooks/useChatSending.ts`; tests: `tests/unit/ask/chat-path-guards.test.tsx` (drives the ACTUAL captured tool executor across POSIX / Windows drive-letter+backslash / UNC roots — allowed-in-workspace reaches the FS with a normalized path; absolute escape, sibling-prefix, case-drift, and `..` all fail closed; `search_files` matter filtering still works on a backslash root). `fileAccessGuards.ts` was already fully on `sameOrInside`, so it needed no change. Verified with a 2-round-clean independent Codex security review. TS-only; no Rust change.
 - **Connector outcome contract (F2.1): no connector sync can fail silently anymore.** Every user-triggered connector sync now ends in exactly one visible outcome — a success with real counts or a plain-language error, never silence — and each leaves a durable, append-only audit row (metadata + counts + a *sanitized error category* only, never raw provider messages, addresses, filenames, or tokens). Brought Mail, Calendly, Box, and Addepar up to the OneDrive/Wealthbox bar:
