+++ src/App.tsx
@@ -5,7 +5,7 @@
  * where AI proposes and the user approves all destructive actions.
  */
 
-import { useState, useCallback, useRef, useEffect, useMemo } from 'react';
+import { useState, useCallback, useRef, useEffect, useMemo, lazy, Suspense } from 'react';
 import { workspacePath } from '@/platform/fs/appPath';
 import { useGlobalEventBus, type AppSurface } from '@/app/lifecycle/useGlobalEventBus';
 import { useAutosave } from '@/app/lifecycle/useAutosave';
@@ -107,15 +107,11 @@ import { useConfirmDialog } from '@/platform/hooks/useConfirmDialog';
 import { ConfirmDialog } from '@/ui/ConfirmDialog';
 import { usePromptDialog } from '@/platform/hooks/usePromptDialog';
 import { useUndoToast } from '@/app/shell/common/UndoToast';
-import { CrmSourcePanel } from '@/features/crm/CrmSourcePanel';
-import { OneDriveSourcePanel } from '@/features/onedrive/OneDriveSourcePanel';
-import { DocusignSourcePanel } from '@/features/docusign/DocusignSourcePanel';
-import { MeetingSourcePanel } from '@/features/calendly/MeetingSourcePanel';
-import { BoxSourcePanel } from '@/features/box/BoxSourcePanel';
-import { SharefileSourcePanel } from '@/features/sharefile/SharefileSourcePanel';
-import { JotformSourcePanel } from '@/features/jotform/JotformSourcePanel';
-import { ZocksSourcePanel } from '@/features/zocks/ZocksSourcePanel';
-import { AddeparSourcePanel } from '@/features/addepar/AddeparSourcePanel';
+
+// Nine connector citation viewers, none of which render anything until their
+// own window event fires — bundled into one lazy chunk (see
+// ConnectorSourcePanels.tsx) so they don't ride into the startup bundle.
+const ConnectorSourcePanels = lazy(() => import('@/app/shell/ConnectorSourcePanels'));
 
 // Module-level constants so the onboarding/tour effects have stable deps
 // and never need to be listed in exhaustive-deps disable comments.
@@ -1463,19 +1459,13 @@ function App() {
         setShowWhatsNewModalDirect={setShowWhatsNewModalDirect}
       />
 
-      {/* Wealthbox CRM citation viewer — listens for lantern:open-crm events
-          dispatched when a `crm:` source link is clicked in a Client Map. */}
-      <CrmSourcePanel />
-      <OneDriveSourcePanel />
-      <DocusignSourcePanel />
-      <MeetingSourcePanel />
-      {/* Bonus connector citation viewers — each listens for its
-          lantern:open-<connector> event dispatched from a Client Map source link. */}
-      <BoxSourcePanel />
-      <SharefileSourcePanel />
-      <JotformSourcePanel />
-      <ZocksSourcePanel />
-      <AddeparSourcePanel />
+      {/* Connector citation viewers (Wealthbox, OneDrive, DocuSign, Calendly,
+          Box, ShareFile, Jotform, Zocks, Addepar) — each self-gates on its own
+          window event and renders nothing until a matching citation is
+          clicked, so a null Suspense fallback is safe here. */}
+      <Suspense fallback={null}>
+        <ConnectorSourcePanels />
+      </Suspense>
     </div>
   );
 }
