+++ src/app/shell/AppDialogs.tsx
@@ -10,6 +10,7 @@
  * local WhatsNewLayer wrapper.
  */
 
+import { lazy, Suspense } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { McpApprovalGate } from '@/features/settings/McpApprovalGate';
@@ -19,7 +20,6 @@ import { MatterManagerDialog } from '@/features/matters/MatterManagerDialog';
 import { InterviewForm } from '@/features/workflows/InterviewForm';
 import { CommandPalette, type PaletteCommand } from '@/app/shell/common/CommandPalette';
 import { SettingsModal } from '@/features/settings/SettingsModal';
-import { AccountWindow } from '@/features/account/AccountWindow';
 import { FeatureTour } from '@/features/onboarding/FeatureTour';
 import { ApiKeyWizard } from '@/features/onboarding/ApiKeyWizard';
 import { ApiKeyManager, type ApiKeyManagerKeychain } from '@/features/settings/ApiKeyManager';
@@ -48,6 +48,12 @@ import { loadAllTemplates } from '@/features/workflows/engine/userTemplates';
 import type { InterviewQuestion } from '@/platform/types/workflow';
 import type { FileNode } from '@/platform/types/workspace';
 import type { SettingCategory } from '@/platform/settings/schema';
+
+// AccountWindow pulls in every connector "Connect" setup panel (Wealthbox,
+// OneDrive, Box, DocuSign, ShareFile, Jotform, Zocks, Addepar, Calendly,
+// Salesforce, Redtail, plus mail + firm admin) — lazy so none of that rides
+// into the startup bundle until the user actually opens Account settings.
+const AccountWindow = lazy(() => import('@/features/account/AccountWindow'));
 
 export interface AppDialogsProps {
   // MCP gate
@@ -259,15 +265,21 @@ export function AppDialogs({
 
       {/* Account / firm window — opened from the rail's account identity, or from
           email connect entry points (pre-selects the Connections tab). */}
-      <AccountWindow
-        open={accountWindowOpen}
-        onOpenChange={(open) => {
-          setAccountWindowOpen(open);
-          if (!open) onAccountWindowClosed?.();
-        }}
-        auditEntries={auditEntries}
-        initialTab={accountWindowInitialTab}
-      />
+      {/* Suspense fallback is `null`, not a spinner: this dialog is mounted
+          unconditionally (Radix Dialog controls visibility via `open`, not
+          mount), so a visible fallback would flash on every app startup while
+          the connector-panel chunk loads in the background. */}
+      <Suspense fallback={null}>
+        <AccountWindow
+          open={accountWindowOpen}
+          onOpenChange={(open) => {
+            setAccountWindowOpen(open);
+            if (!open) onAccountWindowClosed?.();
+          }}
+          auditEntries={auditEntries}
+          initialTab={accountWindowInitialTab}
+        />
+      </Suspense>
 
       {/* Advisor Prep Hero 3.0: rebuilt first-run wizard — the live first-run surface.
           Built above as `firstRunOverlay` so it also renders over the
