+++ src/app/shell/layout/MainPanel.tsx
@@ -74,8 +74,6 @@ import {
 import { FileText, List, PanelRightClose, FileType, X, History, Download, ChevronDown, MoreHorizontal, Columns, Rows, Pencil } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { saveFile } from '@/platform/utils/saveFile';
-import { markdownToDocxBytes } from '@/platform/utils/docx-io';
-import { markdownToPptxBytes } from '@/platform/utils/pptx-io';
 import { withShortcut } from '@/platform/utils/shortcuts';
 import { useConfidentialityMode } from '@/platform/hooks/useConfidentialityMode';
 import { useActiveEgressProvider } from '@/platform/hooks/useActiveEgressProvider';
@@ -1009,6 +1007,7 @@ export function MainPanel({
   const exportAsDocx = useCallback(async () => {
     if (!activeTab) return;
     try {
+      const { markdownToDocxBytes } = await import('@/platform/utils/docx-io');
       const bytes = await markdownToDocxBytes(activeTab.content, activeTab.name);
       const suggestedName = activeTab.name.replace(/\.(md|markdown|txt)$/i, '') + '.docx';
       await saveFile(bytes, {
@@ -1030,6 +1029,7 @@ export function MainPanel({
   const exportAsPptx = useCallback(async () => {
     if (!activeTab) return;
     try {
+      const { markdownToPptxBytes } = await import('@/platform/utils/pptx-io');
       const bytes = await markdownToPptxBytes(activeTab.content);
       const suggestedName = activeTab.name.replace(/\.(md|markdown|txt)$/i, '') + '.pptx';
       await saveFile(bytes, {
